[수정] 테스트코드 수정 및 잘못된 모델 로직수정
