+++ BlackCatTests/FilterTests/MockFilterTask.swift
@@ -1,0 +1,15 @@
+//
+//  MockFilterTask.swift
+//  BlackCatTests
+//
+//  Created by Hamlit Jason on 2022/10/22.
+//
+
+import Foundation
+@testable import BlackCat
+
+//extension FilterTask {
+//    static var mockData0 = [
+//        FilterTask(type: .도안, isSubscribe: true),
+//    
+//}
